requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
torch-cluster>=1.6.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
nuscenes-devkit>=1.1.0
